Upgrade all the things
TL;DR

Upgraded:

- Nextra
- Next
- React
- ESLint
- Codegen and other GraphQL packages

Replaced:

- styled-components and related libraries with plain CSS
- Yarn with pnpm
- Apollo with SWR

Removed:

- Vercel Analytics and Speed Insights
- Unused packages like formik

Changed:

- Homepage layout and fonts to better reflect the new brand guidelines

---

Vercel won't deploy this project anymore:

> Vulnerable version of next-mdx-remote detected (4.4.1). Please update to version 6.0.0 or later.

So, I've upgrade Nextra from v2 to v4. This meant restructuring the content according to the
format it expects: https://the-guild.dev/blog/nextra-4. It also meant upgrading to Next.js 16
and adhering to its app router principles regarding page and layout structuring.
styled-components and Apollo are a bitch to work with in app router, so I've replaced all
styling that used styled-components with CSS Modules, and replaced Apollo data fetching
with useSWR.

I also updated a few styles and fonts to the new brand guidlines, mostly on the homepage.
The docs theme of Nextra is unchanged, so the headings there are not rendered in Yellix.